Added Quiz Functionality
If quiz answer was incorrect, the number of stars that the current level can attain will be reduced by one. Any further mistakes after that will not be counted.

To add:
After the level finishes, call setLevelStars function in levelManager to update value based on the current levels star value.